Split out the work wizard into its own controller refs #1684
