+++ app/views/works_wizard/_wizard_progress.html.erb
