+++ app/views/works_wizard/review.html.erb
@@ -1,7 +1,7 @@
 <div class="wizard-area">
   <h1>New Submission (<%= @work.title %>)</h1>
   <%= render "wizard_progress", wizard_step: 4 %>
-  <%= render "license_text" %>
+  <%= render "/works/license_text" %>
   <p>Data curators will review the submission for accuracy and completeness to make
     it more discoverable and reusable. Recommendations will be available in the "unfinished submissions"
     section of this form within 5-10 business days.</p>
